Require that users provide a given name
We need at least one name so we know who it is, however for many
people a single name will suffice. Treat this as their given
name rather than enforcing it being a first name to provide them
a bit more freedom.

If we were going full https://www.kalzumeus.com/2010/06/17/falsehoods-programmers-believe-about-names/
I'd drop the 'last_name' field and just have the one field, but
this is probably sufficient.